+++ src/representations/msdr/msdrMeasures.cpp
@@ -341,7 +341,7 @@ void msdrMeasure::printShort (ostream& os) const
 
 #ifdef TRACING_IS_ENABLED
   // print the voice measures flat list
-  unsigned int voiceMeasuresFlatListSize =
+  size_t voiceMeasuresFlatListSize =
     fVoiceMeasuresFlatList.size ();
 
   os <<
